all server instances are registered, regarding the structure - go back to the Claude suggestion from Anna
